+++ lib/employees/updateDetails.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'dart:io';
 import 'dart:convert';
 import 'package:image_picker/image_picker.dart';
-import 'package:nannyacademy/nanies/dashboard.dart';
+import 'package:nannyacademy/employees/dashboard.dart';
 
 import 'package:nannyacademy/services/rest_api.dart';
