+++ lib/employers/clientSettings.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:nannyacademy/nanies/dashboard.dart';
-import 'package:nannyacademy/nanies/updateDetails.dart';
+import 'package:nannyacademy/employees/dashboard.dart';
+import 'package:nannyacademy/employees/updateDetails.dart';
+import 'package:nannyacademy/employers/clientPayments.dart';
 
-class Settings extends StatefulWidget {
+class ClientSettings extends StatefulWidget {
   @override
-  _SettingsState createState() => _SettingsState();
+  _ClientSettingsState createState() => _ClientSettingsState();
 }
 
-class _SettingsState extends State<Settings> {
+class _ClientSettingsState extends State<ClientSettings> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -26,14 +27,14 @@ class _SettingsState extends State<Settings> {
         ],
         elevation: 0.0,
         title: Text(
-          'Settings',
+          'Client Settings',
           style: TextStyle(
             fontSize: 20,
             fontFamily: 'Quicksand',
           ),
         ),
         centerTitle: true,
-        backgroundColor: Color.fromRGBO(233, 166, 184, 1),
+        backgroundColor: Color.fromRGBO(255, 200, 124, 1),
       ),
       body: ListView(
         children: <Widget>[
@@ -55,10 +56,15 @@ class _SettingsState extends State<Settings> {
                 ),
               ),
               title: const Text(
-                'Certificates',
+                'Payment Methods',
                 style: TextStyle(color: Colors.black, fontSize: 18.0),
               ),
-              onTap: () => print("ListTile"),
+              onTap: () => Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => ClientPayments(),
+                ),
+              ),
             ),
           ),
           Card(
@@ -72,7 +78,7 @@ class _SettingsState extends State<Settings> {
               contentPadding:
                   EdgeInsets.symmetric(vertical: 10, horizontal: 20),
               leading: CircleAvatar(
-                backgroundColor: Color.fromRGBO(233, 166, 184, 1),
+                backgroundColor: Color.fromRGBO(255, 200, 124, 1),
                 child: Icon(
                   Icons.person,
                   color: Colors.white,
@@ -96,7 +102,7 @@ class _SettingsState extends State<Settings> {
       floatingActionButton: FloatingActionButton(
         tooltip: 'Home',
         elevation: 0.8,
-        backgroundColor: Color.fromRGBO(233, 166, 184, 1),
+        backgroundColor: Color.fromRGBO(255, 200, 124, 1),
         child: const Icon(Icons.home),
         onPressed: () => Navigator.push(
             context, MaterialPageRoute(builder: (context) => Dashboard())),
